fix(caip): keep Node's fs out of the browser-reachable module graph
adapters/coingecko/index.ts re-exports fetchData and parseData from
adapters/coingecko/utils.ts, and utils.ts imported `fs` for writeFiles - a
codegen helper only ever called by generate.ts. That single edge put
`import fs from 'fs'` in the graph every browser bundler reaches from the
package entry, so consuming @shapeshiftoss/caip from a Next.js app fails with
"Module not found: Can't resolve 'fs'".

Move writeFiles into a codegen-only writeFiles.ts per adapter. The public API is
unchanged: fetchCoingeckoData and parseCoingeckoData still export from utils.ts,
and writeFiles was never in a barrel. coincap and coinbase get the same
treatment - their utils.ts is not re-exported today, so the fs import there is
latent rather than live, and this keeps it that way.

Compiled generate.js/writeFiles.js still land in dist, like the compiled tests
already do, but nothing reachable from index.js imports them.
