+++ packages/caip/src/adapters/coincap/utils.test.ts
@@ -6,7 +6,8 @@ import realFs from 'fs'
 import { describe, expect, it, vi } from 'vitest'
 
 import { bitcoinAssetMap, cosmosAssetMap } from '../../utils'
-import { parseData, parseEthData, writeFiles } from './utils'
+import { parseData, parseEthData } from './utils'
+import { writeFiles } from './writeFiles'
 
 const makeEthMockCoincapResponse = () => ({
   id: 'ethereum',
